feat(services): added TMDb and Google Books
