+++ src/Commands/General/manga.ts
@@ -1,19 +1,16 @@
-import {
-	EmbedBuilder,
-	ApplicationCommandOptionType,
-	ChatInputCommandInteraction,
-} from 'discord.js';
+import { EmbedBuilder, ApplicationCommandOptionType, ChatInputCommandInteraction } from 'discord.js';
 import { Discord, Guard, Slash, SlashOption } from 'discordx';
 import { RateLimit, TIME_UNIT } from '@discordx/utilities';
-import { KitsuAPI } from '../../Utils/APIs/kitsuAPI.js';
+import { injectable, inject } from 'tsyringe';
+import { Kitsu } from '@Services';
 
 @Discord()
+@injectable()
 export class Manga {
-	private readonly embed: EmbedBuilder;
+	private embed: EmbedBuilder;
 
-	constructor() {
-		this.embed = new EmbedBuilder().setColor('Blurple').setTimestamp();
-	}
+	// eslint-disable-next-line no-empty-function
+	constructor(@inject(Kitsu) private readonly kitsu: Kitsu) {}
 
 	@Slash({
 		description: 'Get info about a manga using Kitsu.io.',
@@ -35,38 +32,47 @@ export class Manga {
 			title: string,
 			interaction: ChatInputCommandInteraction,
 	) {
-		const kitsu = new KitsuAPI(interaction);
-		const manga = await kitsu.fetchManga(title);
+		this.embed = new EmbedBuilder().setColor('Blurple').setTimestamp();
 
-		return interaction.reply({
-			embeds: [
-				this.embed
-					.setTitle(manga.titles.en_us)
-					.setThumbnail(manga.posterImage.original)
-					.setDescription(manga.synopsis)
-					.addFields(
-						{
-							name: 'Premiered on',
-							value: `<t:${manga.startDateUnix}>`,
-							inline: true,
-						},
-						{
-							name: 'Ended on',
-							value: `<t:${manga.endDateUnix}>`,
-							inline: true,
-						},
-						{
-							name: 'Japanese Title',
-							value: `${manga.titles.ja_JP}`,
-							inline: true,
-						},
-						{
-							name: 'Status',
-							value: manga.status,
-							inline: true,
-						},
+		await this.kitsu.fetchManga(title).then((manga) => {
+			return interaction.reply({
+				embeds: [
+					this.embed
+						.setTitle(manga.titles.en_us)
+						.setThumbnail(manga.posterImage.original)
+						.setDescription(manga.synopsis)
+						.addFields(
+							{
+								name: 'Premiered on',
+								value: `<t:${manga.startDateUnix}>`,
+								inline: true,
+							},
+							{
+								name: 'Ended on',
+								value: `<t:${manga.endDateUnix}>`,
+								inline: true,
+							},
+							{
+								name: 'Japanese Title',
+								value: `${manga.titles.ja_JP}`,
+								inline: true,
+							},
+							{
+								name: 'Status',
+								value: manga.status,
+								inline: true,
+							},
+						),
+				],
+			});
+		}).catch(() => {
+			return interaction.reply({
+				embeds: [
+					this.embed.setDescription(
+						'🔹 | There was an error while fetching the information from the API.',
 					),
-			],
+				],
+			});
 		});
 	}
 }
