+++ src/Commands/General/movie.ts
@@ -1,15 +1,16 @@
-import {
-	EmbedBuilder,
-	ChatInputCommandInteraction,
-	ApplicationCommandOptionType,
-} from 'discord.js';
+import { EmbedBuilder, ChatInputCommandInteraction, ApplicationCommandOptionType } from 'discord.js';
 import { Discord, Slash, SlashOption, Guard } from 'discordx';
-import { TMDBAPI } from '../../Services/tmdbAPI.js';
 import { RateLimit, TIME_UNIT } from '@discordx/utilities';
+import { inject, injectable } from 'tsyringe';
+import { TMDb } from '@Services';
 
 @Discord()
+@injectable()
 export class Movie {
 	private embed: EmbedBuilder;
+
+	// eslint-disable-next-line no-empty-function
+	constructor(@inject(TMDb) private readonly tmdb: TMDb) {}
 
 	@Slash({
 		description: 'Search for a movie using TMDB.',
@@ -31,49 +32,57 @@ export class Movie {
 			title: string,
 			interaction: ChatInputCommandInteraction,
 	) {
-		const API = new TMDBAPI(interaction);
-		const movie = await API.fetchMovie(title);
 		this.embed = new EmbedBuilder().setColor('Blurple').setTimestamp();
 
-		return interaction.reply({
-			embeds: [
-				this.embed
-					.setTitle(movie.title)
-					.setThumbnail(movie.poster)
-					.setDescription(movie.overview)
-					.addFields(
-						{
-							name: 'Genres',
-							value: `> ${movie.genres}`,
-						},
-						{
-							name: 'Production Companies',
-							value: `> ${movie.productionCompanies}`,
-						},
-						{
-							name: 'Released on',
-							value: `> <t:${movie.releaseDateUnix}>`,
-							inline: true,
-						},
-						{
-							name: 'Status',
-							value: `> ${movie.status}`,
-							inline: true,
-						},
-						{
-							name: 'Rating',
-							value: `> ${movie.popularity}/100`,
-							inline: true,
-						},
-						{
-							name: 'Streaming on',
-							value: `> ${movie.streamingServices}`,
-						},
-					)
-					.setFooter({
-						text: 'Information provided by TMDB & JustWatch',
-					}),
-			],
+		await this.tmdb.fetchMovie(title).then((movie) => {
+			return interaction.reply({
+				embeds: [
+					this.embed
+						.setTitle(movie.title)
+						.setThumbnail(movie.poster)
+						.setDescription(movie.overview)
+						.addFields(
+							{
+								name: 'Genres',
+								value: `> ${movie.genres}`,
+							},
+							{
+								name: 'Production Companies',
+								value: `> ${movie.productionCompanies}`,
+							},
+							{
+								name: 'Released on',
+								value: `> <t:${movie.releaseDateUnix}>`,
+								inline: true,
+							},
+							{
+								name: 'Status',
+								value: `> ${movie.status}`,
+								inline: true,
+							},
+							{
+								name: 'Rating',
+								value: `> ${movie.popularity}/100`,
+								inline: true,
+							},
+							{
+								name: 'Streaming on',
+								value: `> ${movie.streamingServices}`,
+							},
+						)
+						.setFooter({
+							text: 'Powered by TMDB & JustWatch',
+						}),
+				],
+			});
+		}).catch(() => {
+			return interaction.reply({
+				embeds: [
+					this.embed.setDescription(
+						'🔹 | There was an error while fetching the information from the API.',
+					),
+				],
+			});
 		});
 	}
 }
