+++ src/Services/index.ts
@@ -1,5 +1,6 @@
 import { RAWG } from './RAWG';
 import { Kitsu } from './Kitsu';
 import { GoogleBooks } from './GoogleBooks';
+import { TMDb } from './TMDb';
 
-export { RAWG, Kitsu, GoogleBooks };
+export { RAWG, Kitsu, GoogleBooks, TMDb };
